게시글 뉴 페이징, 파일 첨부, 삭제

+++ src/com/dao/BoardDao.java
@@ -1,10 +1,16 @@
 package com.dao;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 
 import javax.security.auth.message.callback.PrivateKeyCallback.Request;
+import javax.websocket.Session;
 
 import com.board.beans.Board;
 
@@ -14,6 +20,7 @@ public class BoardDao extends CommonDao {
         return _instance;
     }
 
+//    list1 게시글 불러오기
     public ArrayList<Board> getArticleList() throws SQLException {
         ResultSet rs = null;
 //		인덱스 번호 재조정
@@ -21,7 +28,7 @@ public class BoardDao extends CommonDao {
 //		openConnections().executeQuery(line);
 		
 //      테이블 가져오기  String sql = "select * from BOARD order by idx desc";
-        String sql = "SELECT a.idx, a.title, b.user_id AS writer, a.reg_date, a.hit_count FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx ORDER BY a.idx;";
+        String sql = "SELECT a.idx, a.title, b.user_id AS writer, a.fileName, a.reg_date, a.hit_count FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx ORDER BY a.idx;";
         rs = openConnections().executeQuery(sql);
         
         ArrayList<Board> articleList = new ArrayList<Board>();
@@ -31,6 +38,7 @@ public class BoardDao extends CommonDao {
             article.setIdx(rs.getInt("idx"));
             article.setTitle(rs.getString("title"));
             article.setWriter(rs.getString("writer"));
+            article.setFileName(rs.getString("fileName"));
             article.setRegdate(rs.getString("reg_date"));
             article.setHit_count(rs.getString("hit_count"));
 
@@ -41,6 +49,41 @@ public class BoardDao extends CommonDao {
         return articleList;
     }
     
+//  list2 게시글 불러오기
+    public ArrayList<Board> getArticleList2(int startNum) throws SQLException {
+    	ResultSet rs = null;
+//      게시글 가져오기
+        String sql2 = "SELECT a.idx, a.title, b.user_id AS writer,fileName, fileRealName, a.reg_date, a.hit_count FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx ORDER BY a.idx DESC LIMIT "+startNum+",10;";
+        rs = openConnections().executeQuery(sql2);
+        ArrayList<Board> articleList = new ArrayList<Board>();
+        while (rs.next()) {
+            Board article = new Board();
+
+            article.setIdx(rs.getInt("idx"));
+            article.setTitle(rs.getString("title"));
+            article.setWriter(rs.getString("writer"));
+            article.setFileName(rs.getString("fileName"));
+            article.setFileRealName(rs.getString("fileRealName"));
+            article.setRegdate(rs.getString("reg_date"));
+            article.setHit_count(rs.getString("hit_count"));
+            articleList.add(article);
+        }
+        closeConnections();
+        return articleList;
+    }
+    public int pageNum() throws SQLException {
+    	 ResultSet rs = null;
+//       게시글 수량 확인하기
+         String sql = "SELECT COUNT(idx) AS count FROM BOARD;";
+         rs = openConnections().executeQuery(sql);
+         int count = 0;
+         while(rs.next()) {
+         	count = rs.getInt("count")/10 +1;        	
+         }
+         closeConnections();
+         return count;
+    }
+    
 //    ---------------------------------------------------------------------------------------
 //     게시글 상세보기
     public ArrayList<Board> getArticleList(int idx) throws SQLException {
@@ -49,7 +92,7 @@ public class BoardDao extends CommonDao {
         String sql = "UPDATE my_site.BOARD SET hit_count  = hit_count+1 WHERE idx = "+idx+";";
         rs = openConnections().executeQuery(sql);
         
-        String sql2 = "SELECT a.idx, a.title, b.user_name AS writer, a.reg_date, a.content, a.hit_count FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx WHERE a.idx="+idx+";";
+        String sql2 = "SELECT a.idx, a.title, b.user_id AS writer, a.fileName, a.fileRealName, a.reg_date, a.content, a.hit_count FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx WHERE a.idx="+idx+";";
         rs = openConnections().executeQuery(sql2);
         
         ArrayList<Board> articleList = new ArrayList<Board>();
@@ -59,9 +102,11 @@ public class BoardDao extends CommonDao {
             article.setIdx(rs.getInt("idx"));
             article.setTitle(rs.getString("title"));
             article.setWriter(rs.getString("writer"));
+            article.setFileName(rs.getString("fileName"));
+            article.setFileRealName(rs.getString("fileRealName"));
             article.setRegdate(rs.getString("reg_date"));
             article.setContent(rs.getString("content"));
-            article.setCount(rs.getInt("hit_count"));
+            article.setHit_count(rs.getString("hit_count"));
             articleList.add(article);
         }
         closeConnections();
@@ -70,17 +115,19 @@ public class BoardDao extends CommonDao {
 //    댓글 불러오기
     public ArrayList<Board> getArticleContent(int idx) throws SQLException {
     	ResultSet rs = null;      
-    	String sql = "SELECT idx user_id, text, reg_date FROM CONTENT WHERE num="+idx+";";
+    	String sql = "SELECT idx, user_id, text, reg_date FROM CONTENT WHERE num="+idx+";";
     	rs = openConnections().executeQuery(sql);
     	
     	ArrayList<Board> articleContent = new ArrayList<Board>();
     	while (rs.next()) {
     		Board article = new Board();
     		
+    		article.setIdx(rs.getInt("idx"));
+    		System.out.println(article.getIdx());
     		article.setWriter(rs.getString("user_id"));
+    		System.out.println(article.getWriter());
     		article.setContent(rs.getString("text"));
     		article.setRegdate(rs.getString("reg_date"));
-    		article.setIdx(rs.getInt("idx"));
     		articleContent.add(article);
     	}
     	closeConnections();
@@ -90,14 +137,45 @@ public class BoardDao extends CommonDao {
 
 //  ---------------------------------------------------------------------------------------
 //    게시글 삭제
-    public void deleteArticle(int idx) throws SQLException {
-        String sql = "DELETE FROM BOARD WHERE idx="+idx+";";
-        openConnections().executeQuery(sql);
+    public int deleteArticle(int idx, String id) throws SQLException, IOException {
+    	ResultSet rs = null;
+    	String user_id = "";
+    	String fileName = "";
+    	String sql = "SELECT b.user_id, a.fileRealName FROM BOARD AS a LEFT JOIN USER AS b ON a.user_id = b.idx WHERE a.idx="+idx+";";
+    	rs = openConnections().executeQuery(sql);
+    	while(rs.next()) {
+    		user_id = rs.getString("user_id");    
+    		fileName = rs.getString("fileRealName");
+    	}
+    	closeConnections();
+    	Path path = Paths.get("C:\\Users\\YONSAI\\Desktop\\My\\.metadata\\.plugins\\org.eclipse.wst.server.core\\tmp0\\wtpwebapps\\My_Project\\upload\\"+fileName);
+    	if(user_id.equals(id)) {
+    		Files.delete(path);
+	        String sql2 = "DELETE FROM BOARD WHERE idx="+idx+";";
+	        openConnections().executeQuery(sql2);
+	        return 0;
+    	}else {
+    		return 1;
+    	}
     }
 //    댓글 삭제
-    public void deleteContentArticle(int contentIdx) throws SQLException {
-        String sql = "DELETE FROM CONTENT WHERE idx="+contentIdx+";";
-        openConnections().executeQuery(sql);
+    public int deleteContentArticle(int contentIdx, String id) throws SQLException {
+    	ResultSet rs = null;
+    	String user_id = "";
+    	String sql = "SELECT user_id FROM CONTENT WHERE idx="+contentIdx+";";
+    	rs = openConnections().executeQuery(sql);
+    	while(rs.next()) {
+    		user_id = rs.getString("user_id");    		
+    	}
+    	closeConnections();
+    	
+    	if(user_id.equals(id)) {
+    		String sql2 = "DELETE FROM CONTENT WHERE idx="+contentIdx+";";
+    		openConnections().executeQuery(sql2);
+	        return 0;
+    	}else {
+    		return 1;
+    	}
     }
 
     
@@ -116,8 +194,8 @@ public class BoardDao extends CommonDao {
         
 //    	게시글 작성
     	String sql2 = "INSERT INTO BOARD "
-    			+"(TITLE, hit_count, content, user_id) "
-    			+"VALUES ('"+article.getTitle()+"','"+article.getCount()+"','"+article.getContent()+"','"+idx+"');";
+    			+"(TITLE, hit_count, content,fileName,fileRealName, user_id) "
+    			+"VALUES ('"+article.getTitle()+"','"+article.getCount()+"','"+article.getContent()+"','"+article.getFileName()+"','"+article.getFileRealName()+"','"+idx+"');";
     	openConnections().executeQuery(sql2);
     }
     
@@ -130,4 +208,22 @@ public class BoardDao extends CommonDao {
     	openConnections().executeQuery(sql2);
     }
     
+//  ---------------------------------------------------------------------------------------------
+//    파일첨부
+    public int upload(String fileName, String fileRealName,Board article) throws SQLException {
+    	ResultSet rs = null;
+ 		String sql = "INSERT INTO FILE (fileName,fileRealName, boardNum) VALUES ('"+fileName+"','"+fileRealName+"','"+article.getIdx()+"');";
+
+ 		System.out.println(sql);
+ 		rs = openConnections().executeQuery(sql);
+ 		try {
+ 			PreparedStatement pstmt  = con.prepareStatement(sql);
+ 			pstmt.setString(1,fileName);
+ 			pstmt.setString(2,fileRealName);
+ 			return pstmt.executeUpdate(); 
+ 		} catch (Exception e) {
+ 			// TODO: handle exception
+ 		}
+ 		return -1;
+ 	}
 }
